Structure changes to repo
