+++ tests/test_pylint_datetime.py
@@ -4,7 +4,23 @@
 """Tests for the frequenz.pylint_datetime package."""
 import pytest
 
-from frequenz.lib import delete_me
+
+# TODO(cookiecutter): Remove this function - when it isn't need anymore
+def delete_me(*, blow_up: bool = False) -> bool:
+    """Do stuff for demonstration purposes.
+
+    Args:
+        blow_up: If True, raise an exception.
+
+    Returns:
+        True if no exception was raised.
+
+    Raises:
+        RuntimeError: if blow_up is True.
+    """
+    if blow_up:
+        raise RuntimeError("This function should be removed!")
+    return True
 
 
 def test_pylint_datetime_succeeds() -> None:  # TODO(cookiecutter): Remove
